+++ tests/FsToolkit.ErrorHandling.Tests/AsyncResultCE.fs
@@ -1,9 +1,13 @@
 module AsyncResultCETests
 
 
-#if FABLE_COMPILER
+#if FABLE_COMPILER_PYTHON
+open Fable.Pyxpecto
+#endif
+#if FABLE_COMPILER_JAVASCRIPT
 open Fable.Mocha
-#else
+#endif
+#if !FABLE_COMPILER
 open Expecto
 #endif
 open SampleDomain