+++ tests/FsToolkit.ErrorHandling.Tests/AsyncResultOption.fs
@@ -1,7 +1,11 @@
 module AsyncResultOptionTests
-#if FABLE_COMPILER
+#if FABLE_COMPILER_PYTHON
+open Fable.Pyxpecto
+#endif
+#if FABLE_COMPILER_JAVASCRIPT
 open Fable.Mocha
-#else
+#endif
+#if !FABLE_COMPILER
 open Expecto
 #endif
 open SampleDomain