+++ tests/FsToolkit.ErrorHandling.Tests/List.fs
@@ -1,9 +1,13 @@
 module ListTests
 
 
-#if FABLE_COMPILER
+#if FABLE_COMPILER_PYTHON
+open Fable.Pyxpecto
+#endif
+#if FABLE_COMPILER_JAVASCRIPT
 open Fable.Mocha
-#else
+#endif
+#if !FABLE_COMPILER
 open Expecto
 #endif
 open SampleDomain
@@ -47,6 +51,39 @@ let traverseResultMTests =
                 "traverse the list and return the first error"
     ]
 
+let traverseOptionMTests =
+    testList "List.traverseOptionM Tests" [
+        let tryTweetOption x =
+            match x with
+            | x when String.IsNullOrEmpty x -> None
+            | _ -> Some x
+
+        testCase "traverseOption with a list of valid data"
+        <| fun _ ->
+            let tweets = [
+                "Hi"
+                "Hello"
+                "Hola"
+            ]
+
+            let expected = Some tweets
+            let actual = List.traverseOptionM tryTweetOption tweets
+
+            Expect.equal actual expected "Should have a list of valid tweets"
+
+        testCase "traverseOption with few invalid data"
+        <| fun _ ->
+            let tweets = [
+                "Hi"
+                "Hello"
+                String.Empty
+            ]
+
+            let expected = None
+            let actual = List.traverseOptionM tryTweetOption tweets
+
+            Expect.equal actual expected "traverse the list and return none"
+    ]
 
 let sequenceResultMTests =
     testList "List.sequenceResultM Tests" [
@@ -82,6 +119,38 @@ let sequenceResultMTests =
                 "traverse the list and return the first error"
     ]
 
+let sequenceOptionMTests =
+    testList "List.sequenceOptionM Tests" [
+        let tryTweetOption x =
+            match x with
+            | x when String.IsNullOrEmpty x -> None
+            | _ -> Some x
+
+        testCase "traverseOption with a list of valid data"
+        <| fun _ ->
+            let tweets = [
+                "Hi"
+                "Hello"
+                "Hola"
+            ]
+
+            let expected = Some tweets
+            let actual = List.sequenceOptionM (List.map tryTweetOption tweets)
+
+            Expect.equal actual expected "Should have a list of valid tweets"
+
+        testCase "sequenceOptionM with few invalid data"
+        <| fun _ ->
+            let tweets = [
+                String.Empty
+                "Hello"
+                String.Empty
+            ]
+
+            let actual = List.sequenceOptionM (List.map tryTweetOption tweets)
+
+            Expect.equal actual None "traverse the list and return none"
+    ]
 
 let traverseResultATests =
     testList "List.traverseResultA Tests" [
@@ -285,6 +354,38 @@ let traverseAsyncResultMTests =
         }
     ]
 
+let traverseAsyncOptionMTests =
+
+    let userIds = [
+        userId1
+        userId2
+        userId3
+    ]
+
+    testList "List.traverseAsyncOptionM Tests" [
+        testCaseAsync "traverseAsyncOptionM with a list of valid data"
+        <| async {
+            let expected = Some userIds
+            let f x = async { return Some x }
+            let actual = List.traverseAsyncOptionM f userIds
+
+            match expected with
+            | Some e -> do! Expect.hasAsyncSomeValue e actual
+            | None -> failwith "Error in the test case code"
+        }
+
+        testCaseAsync "traverseOptionA with few invalid data"
+        <| async {
+            let expected = None
+            let f _ = async { return None }
+            let actual = List.traverseAsyncOptionM f userIds
+
+            match expected with
+            | Some _ -> failwith "Error in the test case code"
+            | None -> do! Expect.hasAsyncNoneValue actual
+        }
+    ]
+
 let notifyFailure (PostId _) (UserId uId) =
     async {
         if
@@ -370,6 +471,43 @@ let sequenceAsyncResultMTests =
         }
     ]
 
+let sequenceAsyncOptionMTests =
+
+    let userIds = [
+        userId1
+        userId2
+        userId3
+    ]
+
+    testList "List.sequenceAsyncOptionM Tests" [
+        testCaseAsync "sequenceAsyncOptionM with a list of valid data"
+        <| async {
+            let expected = Some userIds
+            let f x = async { return Some x }
+
+            let actual =
+                List.map f userIds
+                |> List.sequenceAsyncOptionM
+
+            match expected with
+            | Some e -> do! Expect.hasAsyncSomeValue e actual
+            | None -> failwith "Error in the test case code"
+        }
+
+        testCaseAsync "sequenceOptionA with few invalid data"
+        <| async {
+            let expected = None
+            let f _ = async { return None }
+
+            let actual =
+                List.map f userIds
+                |> List.sequenceAsyncOptionM
+
+            match expected with
+            | Some _ -> failwith "Error in the test case code"
+            | None -> do! Expect.hasAsyncNoneValue actual
+        }
+    ]
 
 let sequenceAsyncResultATests =
     let userIds =
@@ -409,16 +547,91 @@ let sequenceAsyncResultATests =
         }
     ]
 
+#if !FABLE_COMPILER
+let traverseVOptionMTests =
+    testList "List.traverseVOptionM Tests" [
+        let tryTweetVOption x =
+            match x with
+            | x when String.IsNullOrEmpty x -> ValueNone
+            | _ -> ValueSome x
+
+        testCase "traverseVOption with a list of valid data"
+        <| fun _ ->
+            let tweets = [
+                "Hi"
+                "Hello"
+                "Hola"
+            ]
+
+            let expected = ValueSome tweets
+            let actual = List.traverseVOptionM tryTweetVOption tweets
+
+            Expect.equal actual expected "Should have a list of valid tweets"
+
+        testCase "traverseVOption with few invalid data"
+        <| fun _ ->
+            let tweets = [
+                "Hi"
+                "Hello"
+                String.Empty
+            ]
+
+            let actual = List.traverseVOptionM tryTweetVOption tweets
+            Expect.equal actual ValueNone "traverse the list and return value none"
+    ]
+
+let sequenceVOptionMTests =
+    testList "List.sequenceVOptionM Tests" [
+        let tryTweetOption x =
+            match x with
+            | x when String.IsNullOrEmpty x -> ValueNone
+            | _ -> ValueSome x
+
+        testCase "traverseVOption with a list of valid data"
+        <| fun _ ->
+            let tweets = [
+                "Hi"
+                "Hello"
+                "Hola"
+            ]
+
+            let expected = ValueSome tweets
+            let actual = List.sequenceVOptionM (List.map tryTweetOption tweets)
+
+            Expect.equal actual expected "Should have a list of valid tweets"
+
+        testCase "sequenceVOptionM with few invalid data"
+        <| fun _ ->
+            let tweets = [
+                String.Empty
+                "Hello"
+                String.Empty
+            ]
+
+            let actual = List.sequenceVOptionM (List.map tryTweetOption tweets)
+            Expect.equal actual ValueNone "traverse the list and return value none"
+    ]
+
+#endif
+
 let allTests =
     testList "List Tests" [
         traverseResultMTests
+        traverseOptionMTests
         sequenceResultMTests
+        sequenceOptionMTests
         traverseResultATests
         sequenceResultATests
         traverseValidationATests
         sequenceValidationATests
         traverseAsyncResultMTests
+        traverseAsyncOptionMTests
         traverseAsyncResultATests
         sequenceAsyncResultMTests
+        sequenceAsyncOptionMTests
         sequenceAsyncResultATests
+#if !FABLE_COMPILER
+        traverseVOptionMTests
+        sequenceVOptionMTests
+#endif
     ]